Section_064_-_Day_64_-_Advanced_-_My_Top_10_Movies_Website/top10movies/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
requests==2.31.0
Flask-WTF==1.2.1
Bootstrap-Flask==2.3.0 
python-dotenv==1.0.0